Remove dead code, fix folder typo, drop unused dependencies
- Fixed Projects import path (Projetcs typo → Projects)
- Deleted typo folder src/pages/Projetcs/
- Removed 5 unused images (blogProject, pizzaProject, eatBurger, Portfoliov2.0, TravisCI-Mascot-2)
- Removed commented-out project cards (Blog, Pizza, Burger) from Projects.js
- Cleaned up empty NavBar left column and dead TravisCI link
- Removed unused dependencies: joi, react-router-dom, webpack
- Added CLAUDE.md with architecture and command reference
