+++ package.json
@@ -5,14 +5,11 @@
   "private": true,
   "dependencies": {
     "gh-pages": "^6.1.1",
-    "joi": "^17.4.2",
     "react": "latest",
     "react-animated-text": "^0.1.4",
     "react-dom": "latest",
-    "react-router-dom": "latest",
     "react-scripts": "^5.0.1",
-    "react-scroll": "latest",
-    "webpack": "latest"
+    "react-scroll": "latest"
   },
   "scripts": {
     "predeploy": "npm run build",
