+++ src/App.js
@@ -3,7 +3,7 @@ import NavBar from "./components/NavBar";
 import Welcome from "./pages/Welcome/Welcome";
 import About from "./pages/About/About";
 import Footer from "./components/Footer";
-import Projects from "./pages/Projetcs/Projects";
+import Projects from "./pages/Projects/Projects";
 import Contact from "./pages/Contact/Contact";
 
 function App() {
